refactor(identity): split errors by layer (domain vs application)
