+++ src/modules/identity/application/use_cases/create_admin_user.py
@@ -4,10 +4,10 @@ from src.modules.identity.application.dtos.identity_dtos import (
     CreateAdminUserInput,
     UserSummary,
 )
+from src.modules.identity.application.errors import UserEmailAlreadyExistsError
 from src.modules.identity.application.ports import PasswordHasherPort
 from src.modules.identity.application.unit_of_work import IdentityUnitOfWorkFactory
 from src.modules.identity.domain.entities.user import User
-from src.modules.identity.domain.errors import UserEmailAlreadyExistsError
 from src.modules.identity.domain.value_objects.email import Email
