+++ src/modules/identity/application/use_cases/get_user_detail.py
@@ -4,9 +4,9 @@ from src.modules.identity.application.dtos.identity_dtos import (
     GetUserDetailInput,
     UserDetail,
 )
+from src.modules.identity.application.errors import UserNotFoundException
 from src.modules.identity.application.unit_of_work import IdentityUnitOfWorkFactory
 from src.modules.identity.application.use_cases._to_output import profile_to_output
-from src.modules.identity.domain.errors import UserNotFoundException
 from src.shared_kernel.value_objects.user_id import UserId
