+++ src/modules/identity/application/use_cases/switch_profile.py
@@ -1,12 +1,12 @@
 """SwitchProfileUseCase."""
 
 from src.modules.identity.application.dtos.identity_dtos import SwitchProfileInput
-from src.modules.identity.application.unit_of_work import IdentityUnitOfWorkFactory
-from src.modules.identity.domain.errors import (
+from src.modules.identity.application.errors import (
     NoActiveSessionError,
     ProfileNotFoundException,
     ProfileOwnershipViolation,
 )
+from src.modules.identity.application.unit_of_work import IdentityUnitOfWorkFactory
 from src.shared_kernel.value_objects.profile_id import ProfileId
 from src.shared_kernel.value_objects.user_id import UserId
