+++ tests/modules/identity/unit/application/use_cases/test_delete_profile.py
@@ -6,16 +6,16 @@ from src.modules.identity.application.dtos.identity_dtos import (
     CreateProfileInput,
     DeleteProfileInput,
 )
+from src.modules.identity.application.errors import (
+    CannotDeleteLastProfileError,
+    ProfileNotFoundException,
+    ProfileOwnershipViolation,
+)
 from src.modules.identity.application.use_cases.create_profile import (
     CreateProfileUseCase,
 )
 from src.modules.identity.application.use_cases.delete_profile import (
     DeleteProfileUseCase,
-)
-from src.modules.identity.domain.errors import (
-    CannotDeleteLastProfileError,
-    ProfileNotFoundException,
-    ProfileOwnershipViolation,
 )
 from src.shared_kernel.value_objects.profile_id import ProfileId
 from src.shared_kernel.value_objects.user_id import UserId
